dataentry-app/pyproject.toml
[tool.poetry]
name = "dataentry"
version = "1.0.0.dev1"
description = "Data Entry"
authors = ["Sébastien VEY <sebvey@gmail.com>"]
repository = "https://github.com/sebvey/marimo-dataentry"
readme = "README.md"
packages = [
    {include = "domain", from = "modules"},
    {include = "checker", from = "modules"},
    {include = "uploader", from = "modules"},
    {include = "pipeline", from = "modules"},
    ]

[tool.poetry.dependencies]
python = "^3.12"
polars = {extras = ["pyarrow"], version = "^0.20.31"}
marimo = "^0.6.21"
python-dateutil = "^2.9.0.post0"
xfp = "^1.0.0b3"

[tool.poetry.group.dev.dependencies]
pyright = "^1.1.368"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
